backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
firebase-admin==6.4.0
gitpython==3.1.37
gunicorn==21.2.0
python-dotenv==1.0.0

